Read Kishou Disaster Prevention Information Feed

Usage

read_kishou_feed(frequency, type)

Arguments

frequency

Release frequency. Select either high frequency ("high") or long term ("low")

type

Feed type. Specify the item to be retrieved as a string. See details for the items.

Value

a tbl object

Details

The following items can be specified in the type argument.

regular: It will be announced on time.

extra: It will be announced at any time.

eqvol: Earthquakes and Volcanoes.

other: Other informations.

See also

Examples

# \donttest{
read_kishou_feed("high", type = "regular")
#> # A tibble: 535 × 6
#>    title                        id      updated             author content link 
#>    <chr>                        <chr>   <dttm>              <chr>  <chr>   <chr>
#>  1 地上実況図                   https:… 2025-01-29 02:02:17 気象庁 【地上実況図… http…
#>  2 大雨危険度通知               https:… 2025-01-29 01:57:24 気象庁 【大雨危険度… http…
#>  3 府県天気概況                 https:… 2025-01-29 01:51:22 前橋地方気… 【天気概況】… http…
#>  4 府県週間天気予報             https:… 2025-01-29 01:50:46 静岡地方気… 【静岡県府県… http…
#>  5 警報級の可能性（明後日以降） https:… 2025-01-29 01:50:46 広島地方気… 【広島県警報… http…
#>  6 警報級の可能性（明後日以降） https:… 2025-01-29 01:50:46 松江地方気… 【島根県警報… http…
#>  7 府県週間天気予報             https:… 2025-01-29 01:50:46 津地方気象… 【三重県府県… http…
#>  8 警報級の可能性（明後日以降） https:… 2025-01-29 01:50:46 彦根地方気… 【滋賀県警報… http…
#>  9 警報級の可能性（明後日以降） https:… 2025-01-29 01:50:46 京都地方気… 【京都府警報… http…
#> 10 警報級の可能性（明後日以降） https:… 2025-01-29 01:50:46 大阪管区気… 【大阪府警報… http…
#> # ℹ 525 more rows
read_kishou_feed("low", "other")
#> # A tibble: 535 × 6
#>    title                  id            updated             author content link 
#>    <chr>                  <chr>         <dttm>              <chr>  <chr>   <chr>
#>  1 季節観測               https://www.… 2025-01-29 00:34:33 宮崎地方気… 【季節観測】… http…
#>  2 季節観測               https://www.… 2025-01-28 23:22:41 熊本地方気… 【季節観測】… http…
#>  3 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:07 沖縄気象台… 【沖縄海上気… http…
#>  4 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:06 福岡管区気… 【長崎海上気… http…
#>  5 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:06 鹿児島地方… 【鹿児島海上… http…
#>  6 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:05 大阪管区気… 【舞鶴海上気… http…
#>  7 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:05 福岡管区気… 【福岡海上気… http…
#>  8 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:04 名古屋地方… 【名古屋海上… http…
#>  9 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:04 高松地方気… 【神戸海上気… http…
#> 10 地方海上予報（Ｈ２８） https://www.… 2025-01-28 21:10:03 気象庁 【東京海上気… http…
#> # ℹ 525 more rows
# }