Read Kishou Disaster Prevention Information Feed

Usage

read_kishou_feed(frequency, type)

Arguments

frequency

Release frequency. Select either high frequency ("high") or long term ("low")

type

Feed type. Specify the item to be retrieved as a string. See details for the items.

Value

a tbl object

Details

The following items can be specified in the type argument.

regular: It will be announced on time.

extra: It will be announced at any time.

eqvol: Earthquakes and Volcanoes.

other: Other informations.

See also

Examples

# \donttest{
read_kishou_feed("high", type = "regular")
#> # A tibble: 370 × 6
#>    title                         id     updated             author content link 
#>    <chr>                         <chr>  <dttm>              <chr>  <chr>   <chr>
#>  1 大雨危険度通知                https… 2024-09-11 05:47:40 気象庁 【大雨… http…
#>  2 大雨危険度通知                https… 2024-09-11 05:37:38 気象庁 【大雨… http…
#>  3 府県天気概況                  https… 2024-09-11 05:34:18 福井…  【天気… http…
#>  4 府県天気予報（Ｒ１）          https… 2024-09-11 05:32:51 福井…  【福井… http…
#>  5 大雨危険度通知                https… 2024-09-11 05:27:38 気象庁 【大雨… http…
#>  6 地上２４時間予想図            https… 2024-09-11 05:26:13 気象庁 【地上… http…
#>  7 アジア太平洋海上悪天２４時間… https… 2024-09-11 05:26:13 気象庁 【アジ… http…
#>  8 大雨危険度通知                https… 2024-09-11 05:17:40 気象庁 【大雨… http…
#>  9 大雨危険度通知                https… 2024-09-11 05:07:45 気象庁 【大雨… http…
#> 10 地上実況図                    https… 2024-09-11 05:05:01 気象庁 【地上… http…
#> # ℹ 360 more rows
read_kishou_feed("low", "other")
#> # A tibble: 418 × 6
#>    title                          id    updated             author content link 
#>    <chr>                          <chr> <dttm>              <chr>  <chr>   <chr>
#>  1 全般海上警報（臨時）           http… 2024-09-11 05:29:58 気象庁 【全般… http…
#>  2 全般海上警報（臨時）（Ｈ２９） http… 2024-09-11 05:29:58 気象庁 【全般… http…
#>  3 生物季節観測                   http… 2024-09-11 02:59:07 熊本…  【生物… http…
#>  4 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 仙台…  【仙台… http…
#>  5 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 新潟…  【新潟… http…
#>  6 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 大阪…  【舞鶴… http…
#>  7 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 福岡…  【福岡… http…
#>  8 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 福岡…  【長崎… http…
#>  9 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 沖縄…  【沖縄… http…
#> 10 地方海上警報（Ｈ２８）         http… 2024-09-11 02:35:27 札幌…  【札幌… http…
#> # ℹ 408 more rows
# }